+++ source/Plugins/PhysX/PhysXVehicleComponent.cpp
@@ -872,7 +872,7 @@ namespace GASS
 			SceneObject::ComponentVector components;
 			GetSceneObject()->GetComponentsByClassName(components,"PhysXBodyComponent");
 
-			for(int i = 0 ; i < components.size(); i++)
+			for(size_t i = 0 ; i < components.size(); i++)
 			{
 				PhysXBodyComponentPtr body = GASS_STATIC_PTR_CAST<PhysXBodyComponent>(components[i]);
 				//if(body.get() != this)
